requirements.txt.txt
streamlit
pandas
yfinance
plotly
openai
google-generativeai